Refactor/graph container (#385)
* break GraphContainer into folder

* remove old GraphContainer file
